undulation bench
add some stuff

fixup

slightly better

add bench

remove casting

Revert "slightly better"

This reverts commit 61f595b71bc86dde274243a6d81aca10a51a50d5.

simplify

+++ egm96/src/egm96_data.rs
@@ -21,7 +21,7 @@
  */
 
 /// Precomputed EGM96 correction and harmonic coefficients
-pub static EGM96_DATA: [[f32; 4]; 65342] = [
+pub static EGM96_DATA: [[f64; 4]; 65342] = [
     [0.0, 0.0, 0.0, 0.0],
     [-5.02745, 0.0, 0.0, 0.0],
     [0.362816, 0.0, 0.0, 0.0],
